add ui

+++ compressor.sh
@@ -1,51 +1,85 @@
 #/bin/sh
 
-#static compressor
+#compressor js
 
-for map in `find ./static -name "*.js.map"`; do
+PWD=`pwd`
+HOME=$PWD
 
-	echo "compressor $map ..."
+if [ -d "./static" ]; then
 
-	dir=${map%/*}
+  	for map in `find ./static -name "*.js.map"`; do
 
-	min=${map%.map}
+		echo "compressor $map ..."
 
-	echo "" > "$min.tmp"
+		dir=${map%/*}
 
-	for js in `cat $map`; do
+		min=${map%.map}
 
-		cat "$dir/$js" >> "$min.tmp"
+		echo "" > "$min.tmp"
 
-	done
+		for js in `cat $map`; do
 
-	java -jar ./yuicompressor.jar --charset utf-8 --type js -o "$min" "$min.tmp"
+			cat "$dir/$js" >> "$min.tmp"
+			echo -e "\n" >> "$min.tmp"
 
-	rm -f "$min.tmp"
+		done
 
-	echo "compressor $map to $min"
+		java -jar $HOME/yuicompressor.jar --charset utf-8 --type js -o "$min" "$min.tmp"
 
-done
+		rm -f "$min.tmp"
 
-for map in `find ./static -name "*.css.map"`; do
-
-	echo "compressor $map ..."
-
-	dir=${map%/*}
-
-	min=${map%.map}
-
-	echo "" > "$min.tmp"
-
-	for js in `cat $map`; do
-
-		cat "$dir/$js" >> "$min.tmp"
+		echo "compressor $map to $min"
 
 	done
 
-	java -jar ./yuicompressor.jar --charset utf-8 --type css -o "$min" "$min.tmp"
+fi
 
-	rm -f "$min.tmp"
-	
-	echo "compressor $map to $min"
-	
-done
+
+
+#compressor css
+
+if [ -d "./static" ]; then
+
+	for map in `find ./static -name "*.css.map"`; do
+
+		echo "compressor $map ..."
+
+		dir=${map%/*}
+
+		min=${map%.map}
+
+		echo "" > "$min.tmp"
+
+		for js in `cat $map`; do
+
+			cat "$dir/$js" >> "$min.tmp"
+			echo -e "\n" >> "$min.tmp"
+
+		done
+
+		java -jar $HOME/yuicompressor.jar --charset utf-8 --type css -o "$min" "$min.tmp"
+
+		rm -f "$min.tmp"
+		
+		echo "compressor $map to $min"
+		
+	done
+
+fi
+
+#compressor html
+
+if [ -d "./@app" ]; then
+
+	for map in `find ./@app -name "*.html.view"`; do
+
+		echo "compressor $map ..."
+
+		go run $HOME/compressor.go -home "$HOME/@app" -o "${map%.view}" -i "$map"
+		
+		echo "compressor $map to ${map%.view}"
+		
+	done
+
+fi
+
